pyproject.toml
[project]
name = "genai-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "databricks-agents>=0.19.0",
    "databricks-connect>=16.1.3",
    "databricks-langchain>=0.4.2",
    "databricks-sdk>=0.50.0",
    "gradio>=5.25.2",
    "ipykernel>=6.29.5",
    "json2html>=1.3.0",
    "langgraph>=0.3.30",
    "markdownify>=1.1.0",
    "mlflow>=2.21.3",
    "notebook>=7.4.0",
    "openai>=1.74.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "unitycatalog-langchain[databricks]>=0.2.0",
]
